add methob post, get, put, delete for order and create security config of User

+++ SHOPAPP-BACKEND/shopapp/shopapp/src/main/java/com/project/shopapp/dtos/OrderDetailDTO.java
@@ -22,7 +22,7 @@ public class OrderDetailDTO {
 
     @Min(value = 0, message = "Price must be greater than or equal to 0")
     @Max(value = 10000000, message = "Price must be less than or equal to 10,000,000")
-    private Long price;
+    private float price;
 
     @JsonProperty("number_of_products")
     @Min(value = 1, message = "number_of_products must be >= 1")
@@ -30,7 +30,7 @@ public class OrderDetailDTO {
 
     @JsonProperty("total_money")
     @Min(value = 0, message = "Total money must be >= 0")
-    private int totalMoney;
+    private float totalMoney;
 
     private String color;
 }
